added formatted amount display to exchange screen
